requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=1.7.6
hydra-core>=1.3.2
omegaconf>=2.3.0
optuna>=3.4.0
shap>=0.44.0
pandera>=0.16.0
matplotlib>=3.7.0
seaborn>=0.12.2
streamlit>=1.28.0
joblib>=1.3
pyyaml>=6.0
pytest>=7.3.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
pre-commit>=3.3.0
